feat: rename agent interface to memory container
