+++ tests/test_cli_ingest_query.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typer.testing import CliRunner
 
 from compact_memory.cli import app
-from compact_memory.agent import Agent
+from compact_memory.memory_container import MemoryContainer
 from compact_memory.vector_store import InMemoryVectorStore
 from compact_memory.embedding_pipeline import get_embedding_dim
 
@@ -26,7 +26,7 @@ def test_ingest_command_removed(tmp_path: Path):
 def test_query_returns_reply(tmp_path: Path, monkeypatch):
     dim = get_embedding_dim()
     store = InMemoryVectorStore(embedding_dim=dim)
-    agent = Agent(store)
+    agent = MemoryContainer(store)
     agent.add_memory("the sky is blue")
     monkeypatch.setattr("compact_memory.cli._load_agent", lambda path: agent)
     result = runner.invoke(app, ["query", "sky?"], env=_env(tmp_path))
